+++ modules/menu.py
@@ -2,8 +2,8 @@ import sys
 
 import pygame
 
-from button import Button
-from state import GameState
+from modules.objects.button import Button
+from modules.state import GameState
 
 BUTTON_WIDTH = 200
 BUTTON_HEIGHT = 50
